Move copy component
